requirements.txt
altgraph==0.17.4
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.3
cx-Freeze==6.15.11
cx-Logging==3.1.0
debugpy==1.6.7
decorator==5.1.1
elastic-transport==8.10.0
elasticsearch==8.11.0
executing==1.2.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.5
lief==0.13.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.26.2
packaging==23.0
pandas==2.1.3
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.14.0
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.0.2
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.6
